Pushed TSH Web Server from Stage Widget to own file
